SSCS-3170 Store the PDF we send to gaps 2 in CCD so we can retrieve it later if it is needed.
Add the pdf we send to Gaps2 to the document store. Then update CCD
with the location so the pdf can be retrieved later if needed. Store
the PDF in the same place a uploaded evidence.
Don't fail if either adding the pdf to the document store or
updating CCD fails.