SSCS-3662 Sending email is not thread safe
The submit you appeal or robotics emails were not being constructed in a
thread safe manor. This could have lead to the emails contain details
from multiple appeals if multiple people are submiting their appeal at
once. Made it so we do not store the email data on the service.