Fix test
Test was failing as we have added a new state where the case moves to
DWP_RESPOND_OVERDUE after the dwp response date. This test had hard
coded dates and was expecting the case to be APPEAL_CREATED. The test
does not really care what the state is so changed the state.